refactor(desktop): move session domain modules to features/session
agent-kind, agent-row-format, session-status, spawn-from-next-action moved.
all importers and vi.mock references updated.
