+++ apps/desktop/src/components/DiffViewerDialog.tsx
@@ -37,7 +37,7 @@ import { openFileInWorkspace } from '../shared/lib/editor';
 import { formatError } from '../shared/lib/errors';
 import { DEFAULT_EDITOR_BINARY, SETTING_DEFAULT_EDITOR, SETTING_EDITOR_BINARY } from '../settings';
 import { useAppStore, useDiffComments, useSummarizerStatus } from '../store';
-import { AGENT_KIND_DEFAULTS } from '../agent-kind';
+import { AGENT_KIND_DEFAULTS } from '../features/session/agent-kind';
 import { STORAGE_KEYS, STORAGE_PREFIXES } from '../shared/lib/storage-keys';
 import {
   listBranchCommits,
