+++ apps/desktop/src/components/NextActionChips.tsx
@@ -5,8 +5,8 @@ import type { NextAction } from '@kay-am/core';
 import type { SessionId } from '@kay-am/types';
 import { formatError } from '../shared/lib/errors';
 import { useAppStore, useSessionNextActions } from '../store';
-import { AGENT_KIND_DEFAULTS, AGENT_KIND_PALETTE } from '../agent-kind';
-import { spawnKindForAction } from '../spawn-from-next-action';
+import { AGENT_KIND_DEFAULTS, AGENT_KIND_PALETTE } from '../features/session/agent-kind';
+import { spawnKindForAction } from '../features/session/spawn-from-next-action';
 
 export interface NextActionChipsProps {
   readonly sessionId: SessionId;
