+++ apps/desktop/src/components/WorkflowNextStepCta.tsx
@@ -3,7 +3,11 @@ import { ArrowRight, ClipboardList } from 'lucide-react';
 import { cn } from '@kay-am/ui';
 import type { Agent, Step, Workflow } from '@kay-am/types';
 import type { VerbosityLevel } from '../verbosity';
-import { AGENT_KIND_DEFAULTS, AGENT_KIND_PALETTE, inferAgentKindFromName } from '../agent-kind';
+import {
+  AGENT_KIND_DEFAULTS,
+  AGENT_KIND_PALETTE,
+  inferAgentKindFromName,
+} from '../features/session/agent-kind';
 
 export interface WorkflowNextStepCtaProps {
   readonly workflow: Workflow;
