+++ apps/desktop/src/features/session/agent-kind.test.ts
