+++ apps/desktop/src/features/session/agent-kind.ts
