+++ apps/desktop/src/features/session/agent-row-format.ts
