+++ apps/desktop/src/features/session/session-status.ts
