+++ apps/desktop/src/features/session/spawn-from-next-action.ts
